# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Mopro Shop API
  version: "1.0.0"
  description: |
    Mobile-first marketplace API for Mopro Shop. Buyer app + seller panel.

    ## Money format
    All monetary values are **integer minor units** (kuruş for TRY; coin-minor for TRY_COIN).
    Every monetary field is accompanied by a `currency` string field.
    Never use decimal values. Display layer divides by 100.

    ## Timestamps
    All timestamps are **RFC 3339 strings** (e.g. `"2026-01-15T10:00:00Z"`).
    Dates (no time component) use ISO 8601 `YYYY-MM-DD`.

    ## Locale resolution
    Product `title` and `description` are resolved server-side from the
    `Accept-Language` request header. The response always contains a single
    resolved string — never a `translations[]` array.
    Fallback chain: requested locale → `tr-TR` → first available locale.

    ## Step-up authentication flow
    Endpoints annotated with `stepUpAuth` require a short-lived step-up token
    (TTL 5 min, claim `scope=high_sensitivity`). Sending a regular access token
    to a `stepUpAuth` endpoint returns `403` with `code: step_up_required`.

    Client flow:
    1. Receive `403 step_up_required`.
    2. Present OTP entry screen.
    3. `POST /auth/step-up` with access token + fresh OTP code → `step_up_token`.
    4. Retry the original request with `Authorization: Bearer <step_up_token>`.

    ## X-Trace-Id
    Clients should generate a UUID per request and attach it as `X-Trace-Id`.
...

    ## Idempotency
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/otp/request:
    post:
      operationId: RequestOtp
      tags: [auth]
      summary: Request a one-time password via SMS
      description: |
        Dispatches a 6-digit OTP to the provided phone number.
        Rate-limited per phone: max 3 requests per 5 minutes.
        X-Idempotency-Key is NOT honored here — each call always dispatches
        a new OTP to avoid replay-suppression attacks.
      security: []
      parameters:
        - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  pattern: '^\+90[5][0-9]{9}$'
                  example: "+905321234567"
                  description: "Turkish mobile number in E.164 format. Must start with +905."
                purpose:
                  type: string
                  enum: [login, step_up]
                  default: login
                  description: |
                    OTP purpose. Use `login` for initial authentication (default).
                    Use `step_up` only if you need a step-up OTP outside the authenticated
                    step-up flow (`POST /auth/step-up/request`). Most clients should omit
                    this field and rely on the default.
      responses:
        '204':
          description: OTP dispatched
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/otp/verify:
    post:
      operationId: VerifyOtp
      tags: [auth]
      summary: Verify OTP and issue access + refresh token pair
      security: []
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone:
                  type: string
                  pattern: '^\+90[5][0-9]{9}$'
                code:
                  type: string
                  pattern: '^\d{6}$'
                  example: "123456"
                purpose:
                  type: string
                  enum: [login, step_up]
                  default: login
                  description: "Must match the purpose used in the corresponding /otp/request call. Defaults to `login`."
      responses:
        '200':
          description: Token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: OTP invalid or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/token/refresh:
    post:
      operationId: RefreshToken
      tags: [auth]
      summary: Exchange a refresh token for a new token pair
      security: []
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New token pair issued; old refresh token revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Refresh token invalid, expired, or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      operationId: Logout
      tags: [auth]
      summary: Revoke the provided refresh token
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '204':
          description: Refresh token revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /me:
    get:
      operationId: GetMe
      tags: [me]
      summary: Get authenticated user profile
      parameters:
        - $ref: '#/components/parameters/TraceId'
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      operationId: UpdateMe
      tags: [me]
      summary: Update user profile fields
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name_first:
                  type: string
                name_last:
                  type: string
                email:
                  type: string
                  format: email
                locale:
                  type: string
                  example: tr-TR
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: DeleteMe
      tags: [me]
      summary: Soft-delete the authenticated user account (KVKK / GDPR)
      description: |
        Requires step-up authentication (`stepUpAuth` security scheme).
        Account enters a 30-day grace period before permanent deletion.
        All active cashback plans are cancelled on confirmation.
      security:
        - stepUpAuth: []
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  enum: [no_longer_needed, privacy_concern, bad_experience, switching_platform, other]
      responses:
        '204':
          description: Account scheduled for deletion
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/StepUpRequired'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /me/devices:
    post:
      operationId: RegisterDevice
      tags: [me]
      summary: Register a device FCM token for push notifications
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fcm_token, device_model, os_version]
              properties:
                fcm_token:
                  type: string
                device_model:
                  type: string
                  example: "Samsung Galaxy S25"
                os_version:
                  type: string
                  example: "Android 16"
      responses:
        '201':
          description: Device registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 access token. TTL 15 min.
        Claims: sub (user_id as string), iat, exp.
        Issued by POST /auth/otp/verify or POST /auth/token/refresh.
    stepUpAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT with claim `scope=high_sensitivity`. TTL 5 min.
        Issued by POST /auth/step-up after re-verifying OTP.
        Required for: DELETE /me (account deletion).
        If a regular bearerAuth token is sent instead, the server returns
        403 with code=step_up_required.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Admin / internal token. Only for POST /products, POST /orders,
        POST /orders/{id}/refund, and GET /categories/{id}/commission.
        Not exposed to mobile clients.
  parameters:
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      description: |
        UUIDv7 generated client-side. Server caches the response for 24 hours
        keyed on this value. Duplicate requests within that window return the
        cached response without re-executing the operation.
      schema:
        type: string
        format: uuid
        example: "01926b7f-1234-7abc-8def-0123456789ab"
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      description: |
        Client-generated trace identifier (UUID or opaque string).
        Echoed in error responses as `error.trace_id`.
        Falls back to a server-generated UUID if absent.
      schema:
        type: string
        example: "4f3a2b1c-e71a-4c3f-b99a-8c3f2a1b7d5e"
  responses:
    BadRequest:
      description: Request is malformed or missing required fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    StepUpRequired:
      description: |
        Regular access token provided but step-up token required.
        Client must call POST /auth/step-up to obtain a step-up token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: step_up_required
              message: "This action requires step-up authentication. Please complete OTP verification."
              trace_id: "4f3a2b1c"
...
    UnprocessableEntity:
      description: Request body failed schema or business-rule validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServiceUnavailable:
      description: |
        System is in read-only mode (maintenance window, migration in progress,
        or circuit breaker open). Retry after the `Retry-After` header value.
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: service_unavailable
              message: "System is temporarily in read-only mode. Please try again shortly."
              trace_id: "4f3a2b1c"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    FieldError:
      type: object
      description: A field-level validation error within an ErrorEnvelope
      required: [name, message]
      properties:
        name:
          type: string
          example: phone
        message:
          type: string
          example: "must start with +905"
    ErrorEnvelope:
      type: object
      description: Standard error response wrapper used by all error status codes
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: string
              description: Machine-readable error slug
              example: validation_error
            message:
              type: string
              description: Human-readable error message (locale from Accept-Language)
              example: "Telefon numarası geçersiz."
            trace_id:
              type: string
              description: Request trace ID. Echoes X-Trace-Id or server-generated UUID.
              example: "4f3a2b1c-e71a-4c3f-b99a-8c3f2a1b7d5e"
            fields:
              type: array
              description: Per-field validation errors. Present only for 422 responses.
              items:
                $ref: '#/components/schemas/FieldError'
...
    TokenPair:
      type: object
      required: [access_token, token_type, expires_in, refresh_token, refresh_expires_at]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: "Seconds until access token expiry. Always 900 (15 min)."
          example: 900
        refresh_token:
          type: string
        refresh_expires_at:
          type: string
          format: date-time
...
    User:
      type: object
      required: [id, phone, locale, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        phone:
          type: string
          example: "+905321234567"
        name_first:
          type: string
          nullable: true
        name_last:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        locale:
          type: string
          example: tr-TR
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Device:
      type: object
      required: [id, fcm_token, device_model, os_version, created_at]
      properties:
        id:
          type: integer
          format: int64
        fcm_token:
          type: string
        device_model:
          type: string
          example: "Samsung Galaxy S25"
        os_version:
          type: string
          example: "Android 16"
        created_at:
          type: string
          format: date-time

    # ── Address ───────────────────────────────────────────────────────────────
...